requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
numpy
tqdm
python-dotenv
requests
